+++ app.js
@@ -1,3 +1,4 @@
+require('./configuration/config')
 var createError = require('http-errors');
 var express = require('express');
 var path = require('path');
@@ -7,18 +8,13 @@ const hbs = require('hbs')
 
 const session = require('express-session');
 const passport = require('passport');
-const LocalStrategy = require('passport-local').Strategy;
 
 const flash = require('connect-flash');
-const mongodb = require('mongodb');
-const mongoose = require('mongoose');
-const expressValidator = require('express-validator');
-const multer = require('multer');
-var upload = multer({
-  dest: './uploads'
-}) //Handle file uploads using ->(multer, 3rd party libr)
+const {
+  mongoose
+} = require('./db/mongoose_config');
 
-var db = mongoose.connection;
+const expressValidator = require('express-validator');
 
 
 var indexRouter = require('./routes/index');
@@ -30,13 +26,15 @@ var app = express();
 app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'hbs');
 hbs.registerPartials(path.join(__dirname + '/views/partials'))
-hbs.localsAsTemplateData(app);//this is for Exposing locals variable as template data
+hbs.localsAsTemplateData(app); //this is for Exposing locals variable as template data
 //when we use app.locals & res.locals -> to set the global variable which can be used in Template engine (Handlebars)
 app.locals.myName = '@Tejas Sabunkar Creations';
 
-app.use(logger('dev'));//morgan's logger
+app.use(logger('dev')); //morgan's logger
 app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
+app.use(express.urlencoded({
+  extended: false
+}));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
@@ -75,21 +73,20 @@ app.use(expressValidator({
 app.use(flash());
 app.use(function (req, res, next) {
   res.locals.messages = require('express-messages')(req, res); //res.locals.messages -> this way we can create a global variable called (messages)
- /* console.log(res.locals.messages());
- console.log(res.locals); */
+  /* console.log(res.locals.messages());
+  console.log(res.locals); */
   next();
 });
 
 //creating a new global variable to hideout login btn when user has logged in
 //and hideout logout btn when user has not loggedin/signin 
-app.get('*', function(req,res,next){//for all pages ->*
- let userObjectFetched = req.user;
- if(!userObjectFetched){ //userObjectFetched in undefined or null
-  res.locals.userLoggedInObj =  null ;
- }
- else{
-  res.locals.userLoggedInObj = userObjectFetched; //created global varaible-> isUserLoggedIn (it is an object type)
- }
+app.get('*', function (req, res, next) { //for all pages ->*
+  let userObjectFetched = req.user;
+  if (!userObjectFetched) { //userObjectFetched in undefined or null
+    res.locals.userLoggedInObj = null;
+  } else {
+    res.locals.userLoggedInObj = userObjectFetched; //created global varaible-> isUserLoggedIn (it is an object type)
+  }
   next();
 });
 
@@ -112,6 +109,6 @@ app.use(function (err, req, res, next) {
   res.render('error');
 });
 
-app.listen(3000);
+// app.listen(3000);
 
 module.exports = app;
