+++ bin/www
@@ -3,7 +3,7 @@
 /**
  * Module dependencies.
  */
-
+require('../configuration/config')
 var app = require('../app');
 var debug = require('debug')('signin:server');
 var http = require('http');
